+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Meet_1_FourWheels_Depot.java
@@ -31,8 +31,10 @@ package org.firstinspires.ftc.teamcode;
 
 import com.disnodeteam.dogecv.CameraViewDisplay;
 import com.disnodeteam.dogecv.DogeCV;
+import com.disnodeteam.dogecv.detectors.roverrukus.GoldAlignDetector;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -43,15 +45,13 @@ public class Meet_1_FourWheels_Depot extends LinearOpMode {
 
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
-    private GoldMineralDetector_2 locator = null;
+    private GoldMineralDetector locator = null;
     //private Lookeebot_4Wheels robot = null;
     private REVTrixbot robot = new REVTrixbot();
 
     private boolean visible = false;
     private boolean done = false;
     private double x = 0.0;
-    private double y = 0.0;
-
     private final static int MIDPOINT = 0;  // screen midpoint
     private final static int LEFTPOINT = -106;
     private final static int RIGHTPOINT = 106;
@@ -76,7 +76,7 @@ public class Meet_1_FourWheels_Depot extends LinearOpMode {
         String text = "??";
 
         // Init Detector
-        locator = new GoldMineralDetector_2();
+        locator = new GoldMineralDetector();
         locator.init(hardwareMap.appContext, CameraViewDisplay.getInstance());
         locator.useDefaults();
 
@@ -85,16 +85,12 @@ public class Meet_1_FourWheels_Depot extends LinearOpMode {
         locator.alignPosOffset = 0; // How far from center frame to offset this alignment zone.
         locator.downscale = 0.4; // How much to downscale the input frames
 
-        //locator.areaScoringMethod = DogeCV.AreaScoringMethod.MAX_AREA; // Can also be PERFECT_AREA
-        locator.areaScoringMethod = DogeCV.AreaScoringMethod.PERFECT_AREA; // Can also be PERFECT_AREA
-        locator.perfectAreaScorer.perfectArea = 2400;  // To be calibrated
+        locator.areaScoringMethod = DogeCV.AreaScoringMethod.MAX_AREA; // Can also be PERFECT_AREA
         //detector.perfectAreaScorer.perfectArea = 10000; // if using PERFECT_AREA scoring
-        //locator.maxAreaScorer.weight = 0.005;
-        locator.perfectAreaScorer.weight = 0.01;
+        locator.maxAreaScorer.weight = 0.005;
 
-        locator.ratioScorer.weight = 50;
-        locator.ratioScorer.perfectRatio = 1.25;  // To be calibrated
-        locator.enable();
+        locator.ratioScorer.weight = 5;
+        locator.ratioScorer.perfectRatio = 1.0;
 
         telemetry.addData("locator", "Initialized");
         telemetry.update();
@@ -102,7 +98,9 @@ public class Meet_1_FourWheels_Depot extends LinearOpMode {
         // Init robot
 
         robot.dt.init(hardwareMap);
+        robot.revTrixbotMineralPaddles.init(hardwareMap);
         robot.idenfierFor5197Depositer.init(hardwareMap);
+        robot.roverRuckusRevTrixBotLift.init(hardwareMap);
 
         // turn on camera
         locator.enable();
@@ -113,29 +111,14 @@ public class Meet_1_FourWheels_Depot extends LinearOpMode {
         waitForStart();
         runtime.reset();
 
-        land();
-        //dropping of the robot
-
         // run until the end of the match (driver presses STOP)
         while (opModeIsActive() && !done) {
         // lineup the camera on the right side
         // right 2 balls are visible
-            sleep(1000);
+            land();
+            sleep(8000);
             visible = locator.isFound();
             x = locator.getXPosition() - MIDPOINT;
-            y = locator.getYPosition();
-
-            if (locator.getArea() < 1200 )
-                visible = false;
-
-            if (locator.getRatio() > 2.5)
-                visible = false;
-
-            if (locator.getScore() > 10)
-                visible = false;
-
-            if (locator.getYPosition() < 120)
-                visible = false;
 
             if(visible) {
                 if (x < 0)
@@ -175,19 +158,33 @@ public class Meet_1_FourWheels_Depot extends LinearOpMode {
             }
             telemetry.addData("IsFound" ,visible); // Is the bot aligned with the gold mineral
             telemetry.addData("X Pos" , x); // Gold X pos.
-            telemetry.addData("Y Pos" , y); // Gold Y pos.
             telemetry.addData("Pos" , text); // Gold X pos.
 
-            telemetry.update();// Update Display
+            telemetry.update();// Gold X pos.
+
+            /*
+            robot.roverRuckusRevTrixBotLift.moveToMinPos(1);
+            */
         }
 
         telemetry.addData("Status" ,"All Done"); // Is the bot aligned with the gold mineral
-        telemetry.update();// Update Display
+        telemetry.update();// Gold X pos.
     }
 
     private void land(){
-        //procedures of the robot landing
+        robot.roverRuckusRevTrixBotLift.setBraking(false);
+        robot.roverRuckusRevTrixBotLift.moveToMaxPos(1);
+        sleep(2000);
+        robot.roverRuckusRevTrixBotLift.setBraking(true);
+        robot.dt.encoderDrive(1, 5, 5);
+        sleep(3000);
+        robot.dt.encoderDrive(1, -20, 20);
+        sleep(1000);
+
+
     }
+
+
 
     private void targetLeft()  {
         // build a profile to handle target on left
