+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/REVTrixbotTankDrive.java
@@ -67,7 +67,7 @@ public class REVTrixbotTankDrive extends ModularRobotIterativeOpMode{
          * The init() method of the hardware class does all the work here
          */
         robot.dt.init(hardwareMap);
-        robot.revTrixbotMineralPaddles.init(hardwareMap);
+        robot.roverRuckusRevTrixBotLift.init(hardwareMap);
 
         //robot.goldLocator.init(hardwareMap); unnecessary
         //telemetry.addData("locator", "Initialized");
@@ -98,7 +98,7 @@ public class REVTrixbotTankDrive extends ModularRobotIterativeOpMode{
     @Override
     public void loop() {
 
-        robot.revTrixbotMineralPaddles.teleOpDeployRetractPaddles(gamepad1.a, gamepad1.b);
+        robot.roverRuckusRevTrixBotLift.teleOpMove(gamepad1.a, gamepad1.b, 1);
 
 
         robot.dt.teleOpTankDrive(gamepad1);
